Add correct and optimize signer batching functions for VRF
We add and explain the signer batching trick lifted from
PrivacyPass.  We could dig more into both the proofs in 3.2.3 of
https://uwspace.uwaterloo.ca/bitstream/handle/10012/8621/Henry_Ryan.pdf
as well as the attack on Peng and Bao's batch proof protocol in
https://www.cypherpunks.ca/~iang/pubs/batchzkp-acns.pdf
We now provide an argument for including the public key however,
so that's likely fine.

We should add a constant time 128 bit scalar multiplication
to curve25519-dalek for some use cases, but the variable time
variant should work for our setting.